requirements.txt
Flask==2.3.3
selenium==4.15.0
gunicorn==21.2.0
python-dotenv==1.0.0